+++ src/main/java/lollipop/Main.java
@@ -1,6 +1,5 @@
 package lollipop;
 
-import lollipop.database.Database;
 import lollipop.listeners.*;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.entities.Activity;
