+++ src/main/java/lollipop/Manager.java
@@ -5,9 +5,9 @@ import lollipop.commands.*;
 import lollipop.commands.duel.Duel;
 import lollipop.commands.duel.Move;
 import lollipop.commands.Eval;
+import lollipop.commands.leaderboard.Leaderboard;
 import lollipop.commands.search.Search;
 import lollipop.commands.trivia.Trivia;
-import lollipop.database.Database;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Guild;
