+++ src/main/java/lollipop/commands/Profile.java
@@ -1,7 +1,7 @@
 package lollipop.commands;
 
 import lollipop.*;
-import lollipop.database.Database;
+import lollipop.Database;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
