+++ src/main/java/lollipop/listeners/PageListener.java
@@ -3,9 +3,6 @@ package lollipop.listeners;
 import awatch.model.Anime;
 import awatch.model.Character;
 import lollipop.API;
-import lollipop.BotStatistics;
-import lollipop.Constant;
-import lollipop.database.Database;
 import lollipop.commands.*;
 import lollipop.commands.Random;
 import lollipop.commands.search.*;
@@ -14,8 +11,6 @@ import lollipop.commands.search.charactercomps.Animes;
 import lollipop.commands.search.charactercomps.Mangas;
 import lollipop.commands.search.charactercomps.VoiceActors;
 import lollipop.commands.search.mangacomps.Chapters;
-import lollipop.commands.trivia.models.TGame;
-import lollipop.commands.trivia.Trivia;
 import lollipop.pages.*;
 import mread.model.Chapter;
 import mread.model.Manga;
@@ -32,7 +27,6 @@ import net.dv8tion.jda.api.interactions.components.buttons.Button;
 import net.dv8tion.jda.api.interactions.components.selections.SelectMenu;
 import org.jetbrains.annotations.NotNull;
 
-import java.awt.*;
 import java.util.*;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
